% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{DCQC}
\alias{DCQC}
\title{Run the DCQC reviewing app}
\usage{
DCQC()
}
\value{
A .rtf of the DCQC review
}
\description{
The SORTEE DCQC reviewing app allows data editors to review papers according to the SORTEE guidelines.
}
